refactor: separate number and attributes of confirmed cases card components
