+++ components/cards/ConfirmedCasesAttributesCard.vue
@@ -1,0 +1,52 @@
+<template>
+  <v-col cols="12" md="6" class="DataCard">
+    <data-table
+      :title="'陽性患者の属性'"
+      :title-id="'attributes-of-confirmed-cases'"
+      :chart-data="patientsTable"
+      :chart-option="{}"
+      :date="Data.patients.date"
+      :info="sumInfoOfPatients"
+      :url="
+        'https://catalog.data.metro.tokyo.lg.jp/dataset/t000010d0000000068'
+      "
+    />
+  </v-col>
+</template>
+
+<script>
+import Data from '@/data/data.json'
+import formatGraph from '@/utils/formatGraph'
+import formatTable from '@/utils/formatTable'
+import DataTable from '@/components/DataTable.vue'
+
+export default {
+  components: {
+    DataTable
+  },
+  data() {
+    // 感染者数グラフ
+    const patientsGraph = formatGraph(Data.patients_summary.data)
+    // 感染者数
+    const patientsTable = formatTable(Data.patients.data)
+
+    const sumInfoOfPatients = {
+      lText: patientsGraph[
+        patientsGraph.length - 1
+      ].cumulative.toLocaleString(),
+      sText: patientsGraph[patientsGraph.length - 1].label + 'の累計',
+      unit: '人'
+    }
+
+    const data = {
+      Data,
+      patientsTable,
+      sumInfoOfPatients,
+    }
+    return data
+  }
+}
+</script>
+
+<style lang="scss">
+</style>
