+++ components/cards/ConfirmedCasesDetailsCard.vue
