+++ components/cards/ConfirmedCasesNumberCard.vue
@@ -1,0 +1,39 @@
+<template>
+  <v-col cols="12" md="6" class="DataCard">
+    <time-bar-chart
+      title="陽性患者数"
+      :title-id="'number-of-confirmed-cases'"
+      :chart-data="patientsGraph"
+      :date="Data.patients.date"
+      :unit="'人'"
+      :url="
+        'https://catalog.data.metro.tokyo.lg.jp/dataset/t000010d0000000068'
+      "
+    />
+  </v-col>
+</template>
+
+<script>
+import Data from '@/data/data.json'
+import formatGraph from '@/utils/formatGraph'
+import TimeBarChart from '@/components/TimeBarChart.vue'
+
+export default {
+  components: {
+    TimeBarChart,
+  },
+  data() {
+    // 感染者数グラフ
+    const patientsGraph = formatGraph(Data.patients_summary.data)
+
+    const data = {
+      Data,
+      patientsGraph,
+    }
+    return data
+  }
+}
+</script>
+
+<style lang="scss">
+</style>
